refactor(ors_control_app): slice RegionBuilder into helpers.ts + Steps.tsx
Begin the RegionBuilder.tsx slice per Phase 4a. Move pure types, constants
and small components into a co-located region-builder/ folder:
- region-builder/helpers.ts: PhaseInfo, GraphInfo, GraphReadiness,
  RegionStatus, ProvisionJob, PROVISION_PHASES, ALL_PROFILES,
  COMPUTE_SIZES, recommendComputeSize, sizeLabel, sizeClass, estTime,
  toCatalogRegion, type aliases.
- region-builder/Steps.tsx: StepsStrip + StepsLegend presentational
  components.

RegionBuilder.tsx is now 1,088 lines (was 1,273) and focuses on the
page-level orchestration. Future commits can split panels (RegionList,
RegionMap, ProvisionForm, BuildProgress) and the polling hook
useProvisionPolling out of the remaining body.

Phase 4a of holistic-control-app-refactor-final, foundation step.
No behaviour change. tsc --noEmit passes cleanly.
